backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
motor==3.3.1
firebase-admin==6.2.0
websockets==11.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
httpx==0.25.0
pydantic-settings==2.0.3
email-validator==2.0.0
orjson==3.9.10
starlette==0.27.0
ratelimit==2.2.1
gunicorn==21.2.0 